+++ CHANGELOG.md
@@ -8,6 +8,7 @@ In this document, Mod is Ctrl on Windows/Linux and Cmd on Mac.
 - Automatically unzip update on Mac.
 - Stop worker threads when engine is paused to save CPU.
 - Hide menu and scrollbars when fullscreen.
+- Add key command (F3) for engine CPU meter.
 - Core
 	- Fix sustain pedal release bug when using polyphonic mode in MIDI-CV.
 - API